requirements.txt
langchain
langchain-huggingface
transformers
requests
python-dotenv
